kindleDisplay/requirements.txt
Flask
requests
pillow
websocket-client
